refactor(fetcher)!: split system fetchers into per-file modules
Splits `src/fetcher/system.rs` (~2900 lines) into a directory module with
one file per fetcher. The omnibus file held 19 fetchers plus shared
helpers, which makes navigation hard and pushes well past the project's
~100-line-per-file guideline.

Layout:
  src/fetcher/system/
    mod.rs              — shared helpers (cached caches, formatters,
                          env detection, payload constructors) + registration
                          entry points + test scaffolding
    dmi.rs              — DMI / SMBIOS reads (was system_dmi.rs)
    info_host.rs        — system_info_host (terminal / os / hostname / ...)
    info_cpu.rs         — system_info_cpu (model / cores / freq / vendor)
    info_memory.rs      — system_info_memory (total / swap_total)
    info_kernel.rs      — system_info_kernel (name / version)
    info_machine.rs     — system_info_machine (DMI host bundle)
    info_board.rs       — system_info_board (DMI motherboard)
    info_bios.rs        — system_info_bios (DMI firmware)
    info_locale.rs      — system_info_locale (single Text)
    info_timezone.rs    — system_info_timezone (single Text)
    info_env.rs         — system_info_env (editor / visual / pager)
    info_desktop.rs     — system_info_desktop (de / wm / init)
    monitor_host.rs     — system_monitor_host (Entries rollup)
    monitor_cpu.rs      — system_monitor_cpu (usage)
    monitor_memory.rs   — system_monitor_memory (usage)
    monitor_uptime.rs   — system_monitor_uptime
    monitor_load.rs     — system_monitor_load
    monitor_processes.rs— system_monitor_processes
    monitor_disk.rs     — system_monitor_disk (cached fetcher)
    monitor_battery.rs  — system_monitor_battery (+ battery helpers)

Each per-fetcher file owns its own `Options` / `Kind` enum / `OPTION_SCHEMAS`
constant; shared helpers (`cached_cpu_info`, `format_*`, `detect_*`,
`parse_options`, `payload`, `entry`, etc.) live in `mod.rs` and are
accessed via `super::*` from submodules. Internal helpers used by the
test module in `mod.rs` are exposed as `pub(crate)` where needed.

`pub use` re-exports every `SystemMonitor*` / `SystemInfo*` struct so
external callers and the registration list see flat names.

Naming preserved from the previous commit on this branch: still
`system_{info|monitor}_{subject}` for every fetcher.
