+++ src/fetcher/system/monitor_uptime.rs
@@ -1,0 +1,38 @@
+//! `system_monitor_uptime` — time since the host last booted.
+
+use sysinfo::System;
+
+use crate::payload::{Body, Payload, TextData};
+use crate::render::Shape;
+use crate::samples;
+
+use super::super::{FetchContext, RealtimeFetcher, Safety};
+use super::{format_uptime, payload};
+
+pub struct SystemMonitorUptime;
+
+impl RealtimeFetcher for SystemMonitorUptime {
+    fn name(&self) -> &str {
+        "system_monitor_uptime"
+    }
+    fn safety(&self) -> Safety {
+        Safety::Safe
+    }
+    fn description(&self) -> &'static str {
+        "Time since the host last booted, formatted as a compact `\"3d 4h\"` / `\"2h 15m\"` / `\"45m\"` string."
+    }
+    fn shapes(&self) -> &[Shape] {
+        &[Shape::Text]
+    }
+    fn sample_body(&self, shape: Shape) -> Option<Body> {
+        match shape {
+            Shape::Text => Some(samples::text("3d 4h 12m")),
+            _ => None,
+        }
+    }
+    fn compute(&self, _ctx: &FetchContext) -> Payload {
+        payload(Body::Text(TextData {
+            value: format_uptime(System::uptime()),
+        }))
+    }
+}
